% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_as_xml.R
\name{expr_as_xml}
\alias{expr_as_xml}
\title{Get an XML representation of an expression}
\usage{
expr_as_xml(expr)
}
\arguments{
\item{An}{expression.}
}
\description{
Get an XML representation of an expression
}
