% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{xml_parse_data}
\alias{xml_parse_data}
\title{Convert R parse data to XML}
\usage{
xml_parse_data(x, includeText = NA, pretty = FALSE)
}
\arguments{
\item{x}{
    an expression returned from \code{\link{parse}}, or a function or other
    object with source reference information
  }

\item{includeText}{
    logical; whether to include the text of parsed items in the result
  }

\item{pretty}{Whether to pretty-indent the XML output.}
}
\value{
An XML string representing the parse data. See details below
}
\description{
TODO
}
\examples{
code <- "function(a = 1, b = 2) {\\n  a + b\\n}\\n"
expr <- parse(text = code, keep.source = TRUE)
cat(xml_parse_data(expr))
}

