% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{xmlparsedata}
\alias{xmlparsedata}
\alias{xmlparsedata-package}
\title{Parse Data of R Code as an 'XML' Tree}
\description{
Convert the output of 'utils::getParseData()' to an 'XML' tree, that is
searchable and easier to manipulate in general.
}

