% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{xml_parse_data}
\alias{xml_parse_data}
\title{Convert R parse data to XML}
\usage{
xml_parse_data(x, includeText = NA, pretty = FALSE, attributes = TRUE)
}
\arguments{
\item{x}{
    an expression returned from \code{\link{parse}}, or a function or other
    object with source reference information
  }

\item{includeText}{
    logical; whether to include the text of parsed items in the result
  }

\item{pretty}{Whether to pretty-indent the XML output. It has a small
overhead which probably only matters for very large source files.}

\item{attributes}{Whether to include location tag attributes (\code{line1}, \code{col1}, ...)}
}
\value{
An XML string representing the parse data. See details below.
}
\description{
In recent R versions the parser can attach source code location
information to the parsed expressions. This information is often
useful for static analysis, e.g. code linting. It can be accessed
via the \code{\link[utils:getParseData]{utils::getParseData()}} function.
}
\details{
\code{xml_parse_data()} converts this information to an XML tree.
The R parser's token names are preserved in the XML as much as
possible, but some of them are not valid XML tag names, so they are
renamed, see the \link{xml_parse_token_map} vector for the
mapping.

The top XML tag is \verb{<exprlist>}, which is a list of
expressions, each expression is an \verb{<expr>} tag. Each tag
has attributes that define the location: \code{line1}, \code{col1},
\code{line2}, \code{col2}. These are from the \code{\link[=getParseData]{getParseData()}}
data frame column names. Next, there are two attributes,
\code{start} and \code{end}, which can be used as an ordering of
expressions in the document. Note that while the values
are correlated with (and in some cases may match exactly)
positions in the document, this cannot be relied upon.

See an example below. See also the README at
\url{https://github.com/r-lib/xmlparsedata#readme}
for examples on how to search the XML tree with the \code{xml2} package
and XPath expressions.

Note that \code{xml_parse_data()} silently drops all control characters
(0x01-0x1f) from the input, except horizontal tab (0x09) and newline
(0x0a), because they are invalid in XML 1.0.
}
\examples{
code <- "function(a = 1, b = 2) {\n  a + b\n}\n"
expr <- parse(text = code, keep.source = TRUE)

# The base R way:
getParseData(expr)

cat(xml_parse_data(expr, pretty = TRUE))
}
\seealso{
\link{xml_parse_token_map} for the token names.
\url{https://github.com/r-lib/xmlparsedata#readme} for more
information and use cases.
}
