% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlparsedata-package.R
\docType{package}
\name{xmlparsedata-package}
\alias{xmlparsedata-package}
\alias{_PACKAGE}
\title{xmlparsedata: Parse Data of 'R' Code as an 'XML' Tree}
\description{
Convert the output of 'utils::getParseData()' to an 'XML' tree, that one can search via 'XPath', and easier to manipulate in general.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/xmlparsedata#readme}
  \item \url{https://r-lib.github.io/xmlparsedata/}
  \item Report bugs at \url{https://github.com/r-lib/xmlparsedata/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Mango Solutions [copyright holder, funder]
}

}
\keyword{internal}
